+++ apps/dev/src/schema/Folder.tsx
@@ -1,7 +1,9 @@
 import {Config, Field} from 'alinea'
+import {Page} from './Page.js'
+import {LinkFields} from './example/LinkFields.js'
 
 export const Folder = Config.type('Folder', {
-  contains: ['Page'],
+  contains: [Page, LinkFields],
   fields: {
     title: Field.text('Title', {
       width: 0.5